fix: rename to jwt to middleware folder
